# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  response.ErrorInfo:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorInfo'
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Cryptocurrency quantitative trading API server
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Crypto Quant API
  version: "1.0"
paths:
...
  /portfolio/pnl/{symbol}:
    get:
      description: Calculate profit/loss for a symbol
      parameters:
      - description: Trading symbol (e.g., BTCUSDT)
        in: path
        name: symbol
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get PnL
      tags:
      - portfolio
  /portfolio/position/{symbol}:
    get:
      description: Get position for a specific symbol
      parameters:
      - description: Trading symbol (e.g., BTCUSDT)
        in: path
        name: symbol
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get position
      tags:
      - portfolio
...
  /wallet/balance/{asset}:
    get:
      description: Get balance for a specific asset
      parameters:
      - description: Asset symbol (e.g., USDT, BTC)
        in: path
        name: asset
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get balance
      tags:
      - wallet
...
swagger: "2.0"